Styling/presentation tweaks
Colorize the header to set it apart, and add a little more spacing to
make everything less busy-looking (we've got plenty of space with this,
so it doesn't matter if we waste some).

I can't figure out how to get the CSS to do the content area so that it
stretches to the bottom of the screen ("viewport") and no further.  So,
it's either force the user to click the actual string to change it, or
have a vertical scrollbar.  The latter seems the lesser of two evils.